TST: extracted F10.7 unit tests
Pulled out F10.7 method unit tests into an appropriately named file.
